+++ util/joystick.h
@@ -6,14 +6,14 @@
 struct Joystick {
     int xId;
     int yId;
-    float deadZone;
-    float sensitivity;
-    float x;
-    float y;
+    double deadZone;
+    double sensitivity;
+    double x;
+    double y;
     bool isXActive;
     bool isYActive;
 
-    Joystick(int xId = -1, int yId = -1, float deadZone = 0, float sensitivity = 0.0, float x = 0.0, float y = 0.0, bool isXActive = false, bool isYActive = false);
+    Joystick(int xId = -1, int yId = -1, double deadZone = 0, double sensitivity = 0.0, double x = 0.0, double y = 0.0, bool isXActive = false, bool isYActive = false);
 };
 
 void initializeJoysticks(const Json::Value &config, Joystick *left = nullptr, Joystick *right = nullptr, Joystick *trigger = nullptr);
